Removed name field from individual PSCs
PSC-54
