+++ src/main/java/uk/gov/companieshouse/pscfiling/api/model/dto/PscDtoCommunal.java
@@ -10,8 +10,6 @@ public interface PscDtoCommunal {
 
     LocalDate getCeasedOn();
 
-    String getName();
-
     List<String> getNaturesOfControl();
 
     LocalDate getNotifiedOn();
